pyproject.toml
[project]
name = "silly-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-litellm>=0.3.5",
    "langchain>=1.2.3",
    "tavily-python>=0.7.17",
    "pydantic>=2.12.5",
    "langgraph>=1.0.5",
    "langchain-xai>=1.2.1",
    "langgraph-cli[inmem]>=0.4.11",
    "python-dotenv>=1.2.1",
]
